201 - move default values classes to a dedicated subfolder
