+++ Kontent.Ai.Management.Tests/ManagementClientTests/ContentTypeTests.cs
@@ -8,6 +8,7 @@ using Kontent.Ai.Management.Tests.Base;
 using Kontent.Ai.Management.Extensions;
 using Kontent.Ai.Management.Models.Shared;
 using Kontent.Ai.Management.Models.Types;
+using Kontent.Ai.Management.Models.Types.Elements.DefaultValues;
 
 namespace Kontent.Ai.Management.Tests.ManagementClientTests;
