+++ Kontent.Ai.Management/Models/Types/Elements/DefaultValues/AssetDefaultValueModel.cs
@@ -1,7 +1,7 @@
 using Kontent.Ai.Management.Models.Shared;
 using System.Collections.Generic;
 
-namespace Kontent.Ai.Management.Models.Types.Elements;
+namespace Kontent.Ai.Management.Models.Types.Elements.DefaultValues;
 
 /// <summary>
 /// The default value model for the asset element
