+++ Kontent.Ai.Management/Models/Types/Elements/DefaultValues/DateElementDefaultValueModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Kontent.Ai.Management.Models.Types.Elements;
+namespace Kontent.Ai.Management.Models.Types.Elements.DefaultValues;
 
 /// <summary>
 /// The default value model for the multiple choice element
