+++ Kontent.Ai.Management/Models/Types/Elements/DefaultValues/NumberElementDefaultValueModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Kontent.Ai.Management.Models.Types.Elements;
+﻿namespace Kontent.Ai.Management.Models.Types.Elements.DefaultValues;
 
 /// <summary>
 /// The default value model for the number element
