+++ Kontent.Ai.Management/Models/Types/Elements/DefaultValues/TextElementDefaultValueModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Kontent.Ai.Management.Models.Types.Elements;
+﻿namespace Kontent.Ai.Management.Models.Types.Elements.DefaultValues;
 
 /// <summary>
 /// The default value model for the text element
