+++ Kontent.Ai.Management/Models/Types/Elements/LinkedItemsElementMetadataModel.cs
@@ -1,4 +1,5 @@
 ﻿using Kontent.Ai.Management.Models.Shared;
+using Kontent.Ai.Management.Models.Types.Elements.DefaultValues;
 using Newtonsoft.Json;
 using System.Collections.Generic;
