+++ Kontent.Ai.Management/Models/Types/Elements/MultipleChoiceElementMetadataModel.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using Kontent.Ai.Management.Models.Types.Elements.DefaultValues;
+using System.Collections.Generic;
 using Newtonsoft.Json;
 
 namespace Kontent.Ai.Management.Models.Types.Elements;
