+++ Kontent.Ai.Management/Models/Types/Elements/TaxonomyElementMetadataModel.cs
@@ -1,4 +1,5 @@
 ﻿using Kontent.Ai.Management.Models.Shared;
+using Kontent.Ai.Management.Models.Types.Elements.DefaultValues;
 using Newtonsoft.Json;
 
 namespace Kontent.Ai.Management.Models.Types.Elements;
